Refactored custom event binders into separate test classes
